refactor: extract diagnostic severity policy
