+++ changelog.d/diagnostic-project-analyzer-policy-decomposition.md
@@ -1,0 +1,5 @@
+---
+category: Changed
+---
+
+- **Changed:** Diagnostic project analysis now delegates effective-severity evaluation, fail-closed analyzer probes, and compiler/generator/analyzer projection to one focused policy collaborator with cross-kind regression coverage. Closes `diagnostic-project-analyzer-policy-decomposition`.
